refactor ltx2 autoencoder tests to use latest mixins
